Extract Credential concern from credential models
