+++ app/models/ai_credential.rb
@@ -1,43 +1,14 @@
-# A user's API credential for one AI provider. Mirrors the AccessToken
-# lifecycle (pending → validating → active|inactive). `credential_data`
-# stores provider-specific fields (e.g. `{ "api_key" => "..." }`) and is
-# encrypted at rest.
+# A user's API credential for one AI provider. `credential_data` stores
+# provider-specific fields (e.g. `{ "api_key" => "..." }`) and is encrypted at
+# rest. Lifecycle, naming, and feed teardown come from Credential; what's here
+# is the model-capability side: which models this key may actually run.
 class AiCredential < ApplicationRecord
-  DISPLAY_NAME_MAX_LENGTH = 80
+  include Credential
+
   REMOVED_EVENT_TYPE = "feed_ai_credential_removed"
-
-  belongs_to :user
-  # `dependent` is handled manually by `disable_dependent_feeds` so we can
-  # both nullify the reference and pull any feed left enabled out of the
-  # enabled state in one pass.
-  has_many :feeds
-
-  # Rails 8 stores the encryption envelope as a JSON object
-  # ({"h": {...}, "p": "<ciphertext>"}) which `jsonb` accepts natively.
-  # The raw column contains only the envelope; the API key is never
-  # stored in plaintext.
-  encrypts :credential_data
-
-  enum :state, { pending: 0, validating: 1, active: 2, inactive: 3 }
+  DEACTIVATED_EVENT_TYPE = "ai_credential_deactivated"
 
   validates :provider, presence: true, inclusion: { in: ->(_) { LlmProvider.names } }
-  validates :display_name,
-            presence: true,
-            length: { maximum: DISPLAY_NAME_MAX_LENGTH },
-            uniqueness: { scope: [:user_id, :provider] }
-
-  validate :api_key_present
-
-  before_validation :assign_name_if_blank, on: :create
-  before_destroy :disable_dependent_feeds
-
-  def default?
-    user.default_ai_credential_id == id
-  end
-
-  def make_default!
-    user.update!(default_ai_credential: self)
-  end
 
   def llm_provider
     LlmProvider.find(provider)
@@ -72,53 +43,6 @@ class AiCredential < ApplicationRecord
   def ruby_llm_context
     RubyLLM.context do |config|
       llm_provider.configure(config, credential_data["api_key"])
-    end
-  end
-
-  def deactivate!(last_error: nil)
-    with_lock do
-      update!(state: :inactive, last_validated_at: Time.current, last_error: last_error)
-      Event.create!(type: "ai_credential_deactivated", level: :warning,
-                    subject: self, user: user)
-      feeds.where(state: Feed.states[:enabled]).update_all(state: Feed.states[:disabled])
-    end
-  end
-
-  private
-
-  def assign_name_if_blank
-    return if display_name.present? || provider.blank?
-
-    self.display_name = generate_unique_name
-  end
-
-  def generate_unique_name
-    label = provider
-    existing = self.class.where(user_id: user_id, provider: provider).pluck(:display_name).map(&:downcase).to_set
-    CredentialNameGenerator.new(label, existing).generate.split.map(&:capitalize).join(" ")
-  end
-
-  def api_key_present
-    return if provider.blank?
-
-    errors.add(:base, "Enter your API key") if credential_data.blank? || credential_data["api_key"].blank?
-  end
-
-  # Detach this credential from every dependent feed. Enabled feeds reuse the
-  # feed's generic disable-and-event transition; drafts and already-disabled
-  # feeds keep their state and receive a removal-only event.
-  def disable_dependent_feeds
-    feeds.find_each do |feed|
-      feed.update_column(:ai_credential_id, nil)
-      next if feed.enabled? && feed.disable_with_event!(REMOVED_EVENT_TYPE, { disabled: true })
-
-      Event.create!(
-        type: REMOVED_EVENT_TYPE,
-        level: :warning,
-        subject: feed,
-        user: user,
-        metadata: { disabled: false }
-      )
     end
   end
 end
